Updated material class
