+++ include/material/interface.hpp
@@ -1,0 +1,8 @@
+#ifndef _MATERIAL_INTERFACE_HPP
+#define _MATERIAL_INTERFACE_HPP
+
+#include "acoustic_material.hpp"
+#include "elastic_material.hpp"
+#include "material.hpp"
+
+#endif
