+++ include/mesh/IO/fortran/read_mesh_database.hpp
@@ -2,7 +2,7 @@
 #define _READ_MESH_DATABASE_HPP
 
 #include "fortran_IO.h"
-#include "material.h"
+#include "kokkos_abstractions.h"
 #include "params.h"
 #include "specfem_mpi.h"
 #include "specfem_setup.hpp"
