+++ src/mesh/IO/fortran/read_mesh_database.cpp
@@ -1,7 +1,6 @@
 #include "mesh/IO/fortran/read_mesh_database.hpp"
 #include "fortran_IO.h"
 #include "kokkos_abstractions.h"
-#include "material.h"
 #include "mesh/IO/fortran/read_material_properties.hpp"
 #include "params.h"
 #include "specfem_mpi.h"
