+++ tests/unittests/mesh/mesh_tests.cpp
@@ -1,6 +1,6 @@
 #include "../Kokkos_Environment.hpp"
 #include "../MPI_environment.hpp"
-#include "material.h"
+#include "material/interface.hpp"
 #include "mesh/mesh.hpp"
 #include "yaml-cpp/yaml.h"
 #include <fstream>
@@ -60,7 +60,7 @@ TEST(MESH_TESTS, fortran_binary_reader) {
   test_config test_config =
       get_test_config(config_filename, MPIEnvironment::mpi_);
 
-  std::vector<specfem::material *> materials;
+  std::vector<specfem::material::material *> materials;
   EXPECT_NO_THROW(specfem::mesh::mesh mesh(test_config.database_filename,
                                            materials, MPIEnvironment::mpi_));
 }
